item status Scheduled doesn't mean Available
I don't know what, if anything, it does mean. But I'm taking out the mapping, so on the off chance we try to migrate an item with this status, the status will go to Folio as Scheduled. And then it will be clear that something needs to be figured out.