Make reference data case insensitive
The capitalization of material types is irregular. There was a minor issue with one material type having different case in Folio than as returned by jrm424.matype() in Oracle. I can't currently think of any reason why case-sensitivity would be necessary for any reference data set, but if it is, this can be made configurable.