Export all items as json
Note that these exported records do not have a holdingRecordId UUID field, but instead a mfhdId field, which should match the hrid on the loaded holdings. In the past, we've populated the holdingRecordId field, but for this initial load of data onto the Ebsco-hosted test instance, we don't know what those identifiers will be.